+++ ownCloud/Client/Actions/Actions+Extensions/DisplayExifMetadataAction.swift
@@ -22,7 +22,7 @@ import ownCloudAppShared
 class DisplayExifMetadataAction : Action {
 	override class var identifier : OCExtensionIdentifier? { return OCExtensionIdentifier("com.owncloud.action.show-exif") }
 	override class var category : ActionCategory? { return .normal }
-	override class var name : String? { return "Image Metadata".localized }
+	override class var name : String? { return "Image metadata".localized }
 	override class var locations : [OCExtensionLocationIdentifier]? { return [.moreItem, .moreDetailItem, .moreFolder, .contextMenuItem] }
 	class var supportedMimeTypes : [String] { return ["image"] }
 	class var excludedMimeTypes : [String] { return ["image/gif", "image/svg"] }