+++ ownCloud/Client/Actions/Actions+Extensions/PresentationModeAction.swift
@@ -34,7 +34,7 @@ class PresentationModeAction: Action {
 	override class func applicablePosition(forContext context: ActionContext) -> ActionPosition {
 		if context.items.first?.cloudStatus == .cloudOnly {
 			return .none
-		} else if let hostViewController = context.viewController, type(of: hostViewController) === DisplayViewController.self {
+		} else if let hostViewController = context.viewController, type(of: hostViewController) === ClientQueryViewController.self {
 			return .none
 		} else if let hostViewController = context.viewController, (hostViewController.navigationController?.isNavigationBarHidden ?? false) {
 			return .none