+++ ownCloud/Client/Viewer/PDF/PDFViewerViewController.swift
@@ -22,16 +22,40 @@ import ownCloudAppShared
 import PDFKit
 import SafariServices
 
-extension UILabel {
-	func _setupPdfInfoLabel() {
+class PulsatingButton: UIButton {
+
+	override func layoutSubviews() {
+		super.layoutSubviews()
+		self.layer.cornerRadius = bounds.size.height / 2
+	}
+
+	override init(frame: CGRect) {
+		super.init(frame: frame)
+		setupShapes()
+	}
+
+	required init?(coder aDecoder: NSCoder) {
+		super.init(coder: aDecoder)
+		setupShapes()
+	}
+
+	fileprivate func setupShapes() {
+		setNeedsLayout()
+		layoutIfNeeded()
+
 		self.layer.masksToBounds = true
-		self.layer.cornerRadius = 8.0
-		self.backgroundColor = UIColor.darkGray
-		self.textColor = UIColor.white
-		self.font = UIFont.systemFont(ofSize: 14.0, weight: UIFont.Weight.medium)
-		self.translatesAutoresizingMaskIntoConstraints = false
-		self.textAlignment = .center
-		self.adjustsFontSizeToFitWidth = true
+		self.backgroundColor = .darkGray
+		self.titleLabel?.textColor = .white
+		self.titleLabel?.font = UIFont.systemFont(ofSize: 14.0, weight: .medium)
+
+		let pulseAnimation = CABasicAnimation(keyPath: "transform.scale")
+		pulseAnimation.duration = 0.7
+		pulseAnimation.fromValue = 1.0
+		pulseAnimation.toValue = 0.85
+		pulseAnimation.timingFunction = CAMediaTimingFunction(name: CAMediaTimingFunctionName.easeInEaseOut)
+		pulseAnimation.autoreverses = true
+		pulseAnimation.repeatCount = 2
+		layer.add(pulseAnimation, forKey: "transformScale")
 	}
 }
 
@@ -57,8 +81,8 @@ class PDFViewerViewController: DisplayViewController, DisplayExtension {
 	private let thumbnailView = PDFThumbnailView()
 
 	private let containerView = UIStackView()
-	private let pageCountLabel = UILabel()
 	private let pageCountContainerView = UIView()
+	private let pageCountButton = PulsatingButton()
 
 	private var searchButtonItem: UIBarButtonItem?
 	private var gotoButtonItem: UIBarButtonItem?
@@ -75,7 +99,7 @@ class PDFViewerViewController: DisplayViewController, DisplayExtension {
 					break
 			}
 
-			pageCountLabel.isHidden = thumbnailViewPosition == .none ? true : false
+			pageCountButton.isHidden = thumbnailViewPosition == .none ? true : false
 
 			setupConstraints()
 		}
@@ -94,7 +118,7 @@ class PDFViewerViewController: DisplayViewController, DisplayExtension {
 		didSet {
 			self.navigationController?.setNavigationBarHidden(fullScreen, animated: true)
 			isFullScreenModeEnabled = fullScreen
-			pageCountLabel.isHidden = fullScreen
+			pageCountButton.isHidden = fullScreen
 			pageCountContainerView.isHidden = fullScreen
 			setupConstraints()
 		}
@@ -106,6 +130,15 @@ class PDFViewerViewController: DisplayViewController, DisplayExtension {
 	static var displayExtensionIdentifier: String = "org.owncloud.pdfViewer.default"
 	static var supportedMimeTypes: [String]? = ["application/pdf", "application/illustrator"]
 	static var features: [String : Any]? = [FeatureKeys.canEdit : false]
+
+	required init() {
+		super.init()
+		updateStrategy = .ask
+	}
+
+	required init?(coder aDecoder: NSCoder) {
+		fatalError("init(coder:) has not been implemented")
+	}
 
 	deinit {
 		NotificationCenter.default.removeObserver(self)
@@ -119,12 +152,10 @@ class PDFViewerViewController: DisplayViewController, DisplayExtension {
 
 	public let searchResultsView = PDFSearchResultsView()
 
-	override func renderSpecificView(completion: @escaping (Bool) -> Void) {
-		if let source = source, let document = PDFDocument(url: source) {
+	override func renderItem(completion: @escaping (Bool) -> Void) {
+		if let source = itemDirectURL, let document = PDFDocument(url: source) {
 			if !didSetupView {
 				didSetupView  = true
-
-				setupToolbar()
 
 				self.thumbnailViewPosition = .none
 
@@ -149,15 +180,18 @@ class PDFViewerViewController: DisplayViewController, DisplayExtension {
 				pdfView.delegate = self
 				containerView.addArrangedSubview(pdfView)
 
+				pageCountButton.translatesAutoresizingMaskIntoConstraints = false
+				pageCountButton.accessibilityLabel = "Go to page".localized
+				pageCountButton.addTarget(self, action: #selector(goToPage), for: .touchDown)
 				pageCountContainerView.translatesAutoresizingMaskIntoConstraints = false
-				pageCountContainerView.addSubview(pageCountLabel)
+				pageCountContainerView.addSubview(pageCountButton)
 
-				pageCountLabel._setupPdfInfoLabel()
+				pageCountButton.centerXAnchor.constraint(equalTo: pageCountContainerView.centerXAnchor).isActive = true
+				pageCountButton.centerYAnchor.constraint(equalTo: pageCountContainerView.centerYAnchor).isActive = true
+				pageCountButton.widthAnchor.constraint(equalTo: pageCountContainerView.widthAnchor, multiplier: 0.25).isActive = true
+				pageCountButton.heightAnchor.constraint(equalTo: pageCountContainerView.heightAnchor, multiplier: 0.5).isActive = true
+				pageCountButton.heightAnchor.constraint(lessThanOrEqualToConstant: 22.0).isActive = true
 
-				pageCountLabel.centerXAnchor.constraint(equalTo: pageCountContainerView.centerXAnchor).isActive = true
-				pageCountLabel.centerYAnchor.constraint(equalTo: pageCountContainerView.centerYAnchor).isActive = true
-				pageCountLabel.widthAnchor.constraint(equalTo: pageCountContainerView.widthAnchor, multiplier: 0.25).isActive = true
-				pageCountLabel.heightAnchor.constraint(equalTo: pageCountContainerView.heightAnchor, multiplier: 0.9).isActive = true
 				containerView.addArrangedSubview(pageCountContainerView)
 
 				self.view.addSubview(containerView)
@@ -194,7 +228,6 @@ class PDFViewerViewController: DisplayViewController, DisplayExtension {
 	}
 
 	// MARK: - View lifecycle management
-
 	override func viewDidLoad() {
 		super.viewDidLoad()
 
@@ -252,7 +285,7 @@ class PDFViewerViewController: DisplayViewController, DisplayExtension {
 	}
 
 	func save(item: OCItem) {
-		if let source = source {
+		if let source = itemDirectURL {
 			editingDelegate?.save(item: item, fileURL: source)
 		}
 	}
@@ -294,7 +327,7 @@ class PDFViewerViewController: DisplayViewController, DisplayExtension {
 		let searchNavigationController = ThemeNavigationController(rootViewController: pdfSearchController)
 		pdfSearchController.pdfDocument = pdfDocument
 		// Interpret the search text and all the matches returned by search view controller
-		pdfSearchController.userSelectedMatchCallback = { (_, matches, selection) in
+		pdfSearchController.userSelectedMatchCallback = { [weak self] (_, matches, selection) in
 			DispatchQueue.main.async { [weak self] in
 				if matches.count > 1 {
 					self?.searchResultsView.matches = matches
@@ -406,19 +439,18 @@ class PDFViewerViewController: DisplayViewController, DisplayExtension {
 
 	}
 
-	private func setupToolbar() {
+	override func composedDisplayBarButtonItems(previous: [UIBarButtonItem]? = nil, itemName: String, itemRemoved: Bool = false) -> [UIBarButtonItem]? {
 		searchButtonItem = UIBarButtonItem(barButtonSystemItem: .search, target: self, action: #selector(search))
-		gotoButtonItem = UIBarButtonItem(image: UIImage(named: "ic_pdf_go_to_page"), style: .plain, target: self, action: #selector(goToPage))
 		outlineItem = UIBarButtonItem(image: UIImage(named: "ic_pdf_outline"), style: .plain, target: self, action: #selector(showOutline))
 
 		searchButtonItem?.accessibilityLabel = "Search PDF".localized
-		gotoButtonItem?.accessibilityLabel = "Go to page".localized
 		outlineItem?.accessibilityLabel = "Outline".localized
 
-		self.parent?.navigationItem.rightBarButtonItems = [
-			gotoButtonItem!,
+		return [
+			actionBarButtonItem,
 			searchButtonItem!,
-			outlineItem!]
+			outlineItem!
+		]
 	}
 
 	// MARK: - Search results navigation
@@ -437,8 +469,8 @@ class PDFViewerViewController: DisplayViewController, DisplayExtension {
 		constraints += horizontal
 		NSLayoutConstraint.activate(constraints)
 
-		self.searchResultsView.updateHandler = { selection in
-			self.jumpTo(selection)
+		self.searchResultsView.updateHandler = { [weak self] selection in
+			self?.jumpTo(selection)
 		}
 
 		self.searchResultsView.closeHandler = { [weak self] in
@@ -473,7 +505,6 @@ class PDFViewerViewController: DisplayViewController, DisplayExtension {
 	}
 
 	// MARK: - Current page selection
-
 	private func selectPage(with label:String) {
 		guard let pdf = pdfView.document else { return }
 
@@ -499,7 +530,8 @@ class PDFViewerViewController: DisplayViewController, DisplayExtension {
 
 		let pageNrText = "\(pdf.index(for: page) + 1)"
 		let maxPageCountText = "\(pdf.pageCount)"
-		pageCountLabel.text = NSString(format: "%@ of %@".localized as NSString, pageNrText, maxPageCountText) as String
+		let title = NSString(format: "%@ of %@".localized as NSString, pageNrText, maxPageCountText) as String
+		pageCountButton.setTitle(title, for: .normal)
 	}
 }